PatentsLandscapes

Documentation for PatentsLandscapes.

PatentsLandscapes.Analysis

PatentsLandscapes.Applicants

PatentsLandscapes.Frequency

PatentsLandscapes.Grouping

PatentsLandscapes.Jurisdictions

PatentsLandscapes.Months

PatentsLandscapes.TaxonomicFilter

PatentsLandscapes.Taxonomy

PatentsLandscapes.TimeResolution

PatentsLandscapes.TimeTrend

PatentsLandscapes.Years

PatentsLandscapes.define_taxon!

PatentsLandscapes.prepdata

PatentsLandscapes.Analysis — Type

Abstract type representing a type of patent analysis.

PatentsLandscapes.Applicants — Type

Dispatch type indicating an analysis of patent data grouped by applicants.

PatentsLandscapes.Frequency — Type

Dispatch type indicating that an analysis of frequencies should be conducted.

PatentsLandscapes.Grouping — Type

Abstract type representing a possible grouping variable in a patent analyis.

PatentsLandscapes.Jurisdictions — Type

Dispatch type indicating an analysis of patent data grouped by jurisdictions.

PatentsLandscapes.Months — Type

Dispatch type indicating a trend analysis should have monthly resolution.

PatentsLandscapes.TaxonomicFilter — Type

Struct representing a database filter using a custom taxonomy.

taxonomy: The name of the taxonomy by which to filter.

included_taxa: The names of the individual taxa to include. If an empty list is passed, all known taxa within the taxonomy are included.

PatentsLandscapes.Taxonomy — Type

Dispatch type indicating an analysis of patent data grouped by a custom taxonomy.

PatentsLandscapes.TimeResolution — Type

Abstract type representing the temporal resolution of a trend analysis.

PatentsLandscapes.TimeTrend — Type

Dispatch type indicating an analysis of patent data over time (trend analysis).

PatentsLandscapes.Years — Type

Dispatch type indicating a trend analysis should have yearly resolution.

PatentsLandscapes.define_taxon! — Method

Define a custom taxonomic group of applications/families, with membership based on an AbstractFilter.

Required arguments:

ds: The data source on which to operate. Membership information for the new taxon is stored in this data source.

taxonomy_name: The name of the taxonomic system, e.g. "base_material".

taxon_name: The name of the taxon or group within the system, e.g. "steel", "wood", etc.

filter: An AbstractFilter that determines which applications or families are included in the taxon.

Keyword arguments:

expand = true: Toggles whether existing membership entries for the taxon are left in place. Otherwise, the taxon is fully redefined based on the filter specified, without respecting existing memberships.

PatentsLandscapes.prepdata — Function

prepdata(
    ds::AbstractDataSource,
    type::Analysis,
    level::AnalysisLevel,
    [filter::AbstractFilter],
    [g1::Grouping],
    [g2::Grouping],
    [g3::Grouping]
)::DataFrame

Perform data retrieval and calculation for an analysis, yielding the results as a DataFrame.

Parameters (in order):

ds: The AbstractDataSource containing the data for the analysis.

type: The type of Analysis to be conducted.

level: The DataLevel for the analysis, e.g. ApplicationLevel(), FamilyLevel() etc.

filter (optional): A filter defining the subset of data to be included in the analysis. If not specified, all data stored in ds is included.

g1 - g3 (optional): Up to three Grouping categories to use in the analysis.